+++ Weekly Workbooks/01 Introduction and Orientation/extras/RedNoise/libs/sdw/OurLine.h
@@ -16,7 +16,7 @@
 bool sortByY(const CanvasPoint& a, const CanvasPoint& b);
 std::vector<float> interpolateSingleFloats(float from, float to, int numberOfValues);
 std::vector<glm::vec3> interpolateThreeElementValues(glm::vec3 from, glm::vec3 to, int numberOfValues);
-std::vector<std::vector<int>> drawLine(DrawingWindow &window, CanvasPoint from, CanvasPoint to, Colour color, std::vector<std::vector<int>> depthMatrix);
+std::vector<std::vector<float>> drawLine(DrawingWindow &window, CanvasPoint from, CanvasPoint to, Colour color, std::vector<std::vector<float>> depthMatrix);
 void drawPoint(DrawingWindow &window, CanvasPoint point, Colour color);
 std::vector<int> unpack(const Colour& color);
 uint32_t pack(std::vector<int> colorgb);
