+++ Weekly Workbooks/01 Introduction and Orientation/extras/RedNoise/libs/sdw/OurTriangle.h
@@ -13,7 +13,7 @@
 #include <string>
 #include "OurLine.h"
 
-std::vector<std::vector<int>> drawStroked(DrawingWindow &window, CanvasTriangle triangle, Colour color, std::vector<std::vector<int>> depthMatrix);
-std::vector<std::vector<int>> drawFilled(DrawingWindow &window, CanvasTriangle triangle, Colour color, std::vector<std::vector<int>> depthMatrix);
+std::vector<std::vector<float>> drawStroked(DrawingWindow &window, CanvasTriangle triangle, Colour color, std::vector<std::vector<float>> depthMatrix);
+std::vector<std::vector<float>> drawFilled(DrawingWindow &window, CanvasTriangle triangle, Colour color, std::vector<std::vector<float>> depthMatrix);
 CanvasTriangle randomTriangle();
 Colour randomColor();
